requirements.txt
altair==5.5.0
appdirs==1.4.4
asttokens==3.0.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.6.1
contourpy==1.3.1
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
executing==2.1.0
fake-useragent==2.0.3
feedparser==6.0.11
fonttools==4.55.3
frozendict==2.4.6
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka==1.3.5
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
markdownify==0.14.1
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.11
narwhals==1.17.0
nest-asyncio==1.6.0
numpy==2.2.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parse==1.20.2
parso==0.8.4
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
pyChatGPT==0.4.3.3
pycparser==2.22
pydeck==0.9.1
pyee==11.1.1
Pygments==2.18.0
pyparsing==3.2.0
pyppeteer==2.0.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-html==0.10.0
rich==13.9.4
rpds-py==0.22.3
selenium==4.27.1
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
streamlit==1.41.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
urllib3==1.26.20
w3lib==2.2.1
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
yfinance==0.2.50
zipp==3.21.0
